+++ app/src/main/scala/com/waz/zclient/calling/CallParticipantsFragment.scala
@@ -18,18 +18,47 @@
 package com.waz.zclient.calling
 
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView.ViewHolder
-import android.support.v7.widget.{RecyclerView, Toolbar}
+import android.support.v7.widget.Toolbar
 import android.view.{LayoutInflater, View, ViewGroup}
 import com.waz.ZLog.ImplicitTag._
-import com.waz.zclient.{FragmentHelper, R}
+import com.waz.utils.events.Subscription
+import com.waz.zclient.R
+import com.waz.zclient.calling.controllers.CallController
+import com.waz.zclient.calling.views.CallParticipantsView
+import com.waz.zclient.utils.RichView
 
-class CallParticipantsFragment extends FragmentHelper {
+class CallParticipantsFragment extends FadingControls {
+  private lazy val toolbar = view[Toolbar](R.id.toolbar)
+  private lazy val participantsView = view[CallParticipantsView](R.id.full_call_participants)
 
-  val toolbar = view[Toolbar](R.id.toolbar)
+  private lazy val controller = inject[CallController]
+
+  private var subs = Set.empty[Subscription]
 
   override def onCreateView(inflater: LayoutInflater, container: ViewGroup, savedInstanceState: Bundle) =
     inflater.inflate(R.layout.fragment_calling_participants, container, false)
+
+  override def onViewCreated(view: View, savedInstanceState: Bundle): Unit = {
+    super.onViewCreated(view, savedInstanceState)
+
+    view.onClick(if (controller.showVideoView.currentValue.getOrElse(false)) toggleControlVisibility())
+
+    toolbar
+    participantsView
+
+    setFadingControls(toolbar, participantsView)
+
+    subs += controller.showVideoView.onUi {
+      case true  => extendControlsDisplay()
+      case false => stopFadeOut()
+    }
+  }
+
+  override def onDestroyView(): Unit = {
+    subs.foreach(_.destroy())
+    subs = Set.empty
+    super.onDestroyView()
+  }
 
   override def onResume(): Unit = {
     super.onResume()
@@ -43,22 +72,9 @@ class CallParticipantsFragment extends FragmentHelper {
     toolbar.foreach(_.setNavigationOnClickListener(null))
     super.onPause()
   }
-
 }
 
 object CallParticipantsFragment {
-
-  class AllParticipantsAdapter() extends RecyclerView.Adapter[ViewHolder] {
-
-
-    override def getItemCount = ???
-
-    override def onCreateViewHolder(parent: ViewGroup, viewType: Int) = ???
-
-    override def onBindViewHolder(holder: ViewHolder, position: Int) = ???
-  }
-
-
 
   def apply(): CallParticipantsFragment = new CallParticipantsFragment()
